Move MAD and quantile states/operations to separate files
