+++ src/core_functions/aggregate/holistic/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library_unity(duckdb_aggr_holistic OBJECT quantile.cpp mode.cpp
+add_library_unity(duckdb_aggr_holistic OBJECT quantile.cpp mad.cpp mode.cpp
                   approximate_quantile.cpp reservoir_quantile.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_aggr_holistic>
