+++ src/core_functions/aggregate/holistic/quantile.cpp
@@ -4,101 +4,14 @@
 #include "duckdb/planner/expression.hpp"
 #include "duckdb/common/operator/cast_operators.hpp"
 #include "duckdb/common/operator/abs.hpp"
-#include "duckdb/common/operator/multiply.hpp"
-#include "duckdb/core_functions/aggregate/quantile_helpers.hpp"
-#include "duckdb/core_functions/aggregate/quantile_sort_tree.hpp"
+#include "duckdb/core_functions/aggregate/quantile_state.hpp"
 #include "duckdb/common/types/timestamp.hpp"
 #include "duckdb/common/queue.hpp"
 #include "duckdb/common/serializer/serializer.hpp"
 #include "duckdb/common/serializer/deserializer.hpp"
-
 #include "SkipList.h"
 
 namespace duckdb {
-
-struct FrameSet {
-	inline explicit FrameSet(const SubFrames &frames_p) : frames(frames_p) {
-	}
-
-	inline idx_t Size() const {
-		idx_t result = 0;
-		for (const auto &frame : frames) {
-			result += frame.end - frame.start;
-		}
-
-		return result;
-	}
-
-	inline bool Contains(idx_t i) const {
-		for (idx_t f = 0; f < frames.size(); ++f) {
-			const auto &frame = frames[f];
-			if (frame.start <= i && i < frame.end) {
-				return true;
-			}
-		}
-		return false;
-	}
-	const SubFrames &frames;
-};
-
-struct QuantileReuseUpdater {
-	idx_t *index;
-	idx_t j;
-
-	inline QuantileReuseUpdater(idx_t *index, idx_t j) : index(index), j(j) {
-	}
-
-	inline void Neither(idx_t begin, idx_t end) {
-	}
-
-	inline void Left(idx_t begin, idx_t end) {
-	}
-
-	inline void Right(idx_t begin, idx_t end) {
-		for (; begin < end; ++begin) {
-			index[j++] = begin;
-		}
-	}
-
-	inline void Both(idx_t begin, idx_t end) {
-	}
-};
-
-void ReuseIndexes(idx_t *index, const SubFrames &currs, const SubFrames &prevs) {
-
-	//  Copy overlapping indices by scanning the previous set and copying down into holes.
-	//	We copy instead of leaving gaps in case there are fewer values in the current frame.
-	FrameSet prev_set(prevs);
-	FrameSet curr_set(currs);
-	const auto prev_count = prev_set.Size();
-	idx_t j = 0;
-	for (idx_t p = 0; p < prev_count; ++p) {
-		auto idx = index[p];
-
-		//  Shift down into any hole
-		if (j != p) {
-			index[j] = idx;
-		}
-
-		//  Skip overlapping values
-		if (curr_set.Contains(idx)) {
-			++j;
-		}
-	}
-
-	//  Insert new indices
-	if (j > 0) {
-		QuantileReuseUpdater updater(index, j);
-		AggregateExecutor::IntersectFrames(prevs, currs, updater);
-	} else {
-		//  No overlap: overwrite with new values
-		for (const auto &curr : currs) {
-			for (auto idx = curr.start; idx < curr.end; ++idx) {
-				index[j++] = idx;
-			}
-		}
-	}
-}
 
 template <class INPUT_TYPE>
 struct IndirectLess {
@@ -328,267 +241,6 @@ string_t CastInterpolation::Cast(const string_t &src, Vector &result) {
 //===--------------------------------------------------------------------===//
 // Quantile State
 //===--------------------------------------------------------------------===//
-template <class T>
-struct PointerLess {
-	inline bool operator()(const T &lhi, const T &rhi) const {
-		return *lhi < *rhi;
-	}
-};
-
-template<typename INPUT_TYPE>
-struct WindowQuantileState {
-	// Windowed Quantile merge sort trees
-	using QuantileSortTree32 = QuantileSortTree<uint32_t>;
-	using QuantileSortTree64 = QuantileSortTree<uint64_t>;
-	unique_ptr<QuantileSortTree32> qst32;
-	unique_ptr<QuantileSortTree64> qst64;
-
-	// Windowed Quantile skip lists
-	using PointerType = const INPUT_TYPE *;
-	using SkipListType = duckdb_skiplistlib::skip_list::HeadNode<PointerType, PointerLess<PointerType>>;
-	SubFrames prevs;
-	unique_ptr<SkipListType> s;
-	mutable vector<PointerType> dest;
-
-	// Windowed MAD indirection
-	idx_t count;
-	vector<idx_t> m;
-
-	WindowQuantileState() : count(0) {}
-
-	inline void SetCount(size_t count_p) {
-		count = count_p;
-		if (count >= m.size()) {
-			m.resize(count);
-		}
-	}
-
-	inline SkipListType &GetSkipList(bool reset = false) {
-		if (reset || !s) {
-			s.reset();
-			s = make_uniq<SkipListType>();
-		}
-		return *s;
-	}
-
-	struct SkipListUpdater {
-		SkipListType &skip;
-		const INPUT_TYPE *data;
-		const QuantileIncluded &included;
-
-		inline SkipListUpdater(SkipListType &skip, const INPUT_TYPE *data, const QuantileIncluded &included)
-		    : skip(skip), data(data), included(included) {
-		}
-
-		inline void Neither(idx_t begin, idx_t end) {
-		}
-
-		inline void Left(idx_t begin, idx_t end) {
-			for (; begin < end; ++begin) {
-				if (included(begin)) {
-					skip.remove(data + begin);
-				}
-			}
-		}
-
-		inline void Right(idx_t begin, idx_t end) {
-			for (; begin < end; ++begin) {
-				if (included(begin)) {
-					skip.insert(data + begin);
-				}
-			}
-		}
-
-		inline void Both(idx_t begin, idx_t end) {
-		}
-	};
-
-	void UpdateSkip(const INPUT_TYPE *data, const SubFrames &frames, const QuantileIncluded &included) {
-		//	No overlap, or no data
-		if (!s || prevs.back().end <= frames.front().start || frames.back().end <= prevs.front().start) {
-			auto &skip = GetSkipList(true);
-			for (const auto &frame : frames) {
-				for (auto i = frame.start; i < frame.end; ++i) {
-					if (included(i)) {
-						skip.insert(data + i);
-					}
-				}
-			}
-		} else {
-			auto &skip = GetSkipList();
-			SkipListUpdater updater(skip, data, included);
-			AggregateExecutor::IntersectFrames(prevs, frames, updater);
-		}
-	}
-
-	bool HasTrees() const {
-		return qst32 || qst64;
-	}
-
-	template <typename RESULT_TYPE, bool DISCRETE>
-	RESULT_TYPE WindowScalar(const INPUT_TYPE *data, const SubFrames &frames, const idx_t n, Vector &result,
-	                         const QuantileValue &q) const {
-		D_ASSERT(n > 0);
-		if (qst32) {
-			return qst32->WindowScalar<INPUT_TYPE, RESULT_TYPE, DISCRETE>(data, frames, n, result, q);
-		} else if (qst64) {
-			return qst64->WindowScalar<INPUT_TYPE, RESULT_TYPE, DISCRETE>(data, frames, n, result, q);
-		} else if (s) {
-			// Find the position(s) needed
-			try {
-				Interpolator<DISCRETE> interp(q, s->size(), false);
-				s->at(interp.FRN, interp.CRN - interp.FRN + 1, dest);
-				return interp.template Extract<INPUT_TYPE, RESULT_TYPE>(dest.data(), result);
-			} catch (const duckdb_skiplistlib::skip_list::IndexError &idx_err) {
-				throw InternalException(idx_err.message());
-			}
-		} else {
-			throw InternalException("No accelerator for scalar QUANTILE");
-		}
-	}
-
-	template <typename CHILD_TYPE, bool DISCRETE>
-	void WindowList(const INPUT_TYPE *data, const SubFrames &frames, const idx_t n, Vector &list, const idx_t lidx,
-	                const QuantileBindData &bind_data) const {
-		D_ASSERT(n > 0);
-		// Result is a constant LIST<CHILD_TYPE> with a fixed length
-		auto ldata = FlatVector::GetData<list_entry_t>(list);
-		auto &lentry = ldata[lidx];
-		lentry.offset = ListVector::GetListSize(list);
-		lentry.length = bind_data.quantiles.size();
-
-		ListVector::Reserve(list, lentry.offset + lentry.length);
-		ListVector::SetListSize(list, lentry.offset + lentry.length);
-		auto &result = ListVector::GetEntry(list);
-		auto rdata = FlatVector::GetData<CHILD_TYPE>(result);
-
-		for (const auto &q : bind_data.order) {
-			const auto &quantile = bind_data.quantiles[q];
-			rdata[lentry.offset + q] = WindowScalar<CHILD_TYPE, DISCRETE>(data, frames, n, result, quantile);
-		}
-	}
-};
-
-template <typename INPUT_TYPE, typename SAVE_TYPE>
-struct QuantileState {
-	using SaveType = SAVE_TYPE;
-	using InputType = INPUT_TYPE;
-
-	// Regular aggregation
-	vector<SaveType> v;
-
-	// Window Quantile State
-	unique_ptr<WindowQuantileState<INPUT_TYPE>> window_state;
-
-	bool HasTrees() const {
-		return window_state && window_state->HasTrees();
-	}
-	WindowQuantileState<INPUT_TYPE> &GetOrCreateWindowState() {
-		if (!window_state) {
-			window_state = make_uniq<WindowQuantileState<INPUT_TYPE>>();
-		}
-		return *window_state;
-	}
-	WindowQuantileState<INPUT_TYPE> &GetWindowState() {
-		return *window_state;
-	}
-	const WindowQuantileState<INPUT_TYPE> &GetWindowState() const {
-		return *window_state;
-	}
-};
-
-struct QuantileOperation {
-	template <class STATE>
-	static void Initialize(STATE &state) {
-		new (&state) STATE();
-	}
-
-	template <class INPUT_TYPE, class STATE, class OP>
-	static void ConstantOperation(STATE &state, const INPUT_TYPE &input, AggregateUnaryInput &unary_input,
-	                              idx_t count) {
-		for (idx_t i = 0; i < count; i++) {
-			Operation<INPUT_TYPE, STATE, OP>(state, input, unary_input);
-		}
-	}
-
-	template <class INPUT_TYPE, class STATE, class OP>
-	static void Operation(STATE &state, const INPUT_TYPE &input, AggregateUnaryInput &) {
-		state.v.emplace_back(input);
-	}
-
-	template <class STATE, class OP>
-	static void Combine(const STATE &source, STATE &target, AggregateInputData &) {
-		if (source.v.empty()) {
-			return;
-		}
-		target.v.insert(target.v.end(), source.v.begin(), source.v.end());
-	}
-
-	template <class STATE>
-	static void Destroy(STATE &state, AggregateInputData &) {
-		state.~STATE();
-	}
-
-	static bool IgnoreNull() {
-		return true;
-	}
-
-	template <class STATE, class INPUT_TYPE>
-	static void WindowInit(AggregateInputData &aggr_input_data, const WindowPartitionInput &partition,
-	                       data_ptr_t g_state) {
-		D_ASSERT(partition.input_count == 1);
-
-		auto inputs = partition.inputs;
-		const auto count = partition.count;
-		const auto &filter_mask = partition.filter_mask;
-		const auto &stats = partition.stats;
-
-		//	If frames overlap significantly, then use local skip lists.
-		if (stats[0].end <= stats[1].begin) {
-			//	Frames can overlap
-			const auto overlap = double(stats[1].begin - stats[0].end);
-			const auto cover = double(stats[1].end - stats[0].begin);
-			const auto ratio = overlap / cover;
-			if (ratio > .75) {
-				return;
-			}
-		}
-
-		const auto data = FlatVector::GetData<const INPUT_TYPE>(inputs[0]);
-		const auto &data_mask = FlatVector::Validity(inputs[0]);
-
-		//	Build the tree
-		auto &state = *reinterpret_cast<STATE *>(g_state);
-		auto &window_state = state.GetOrCreateWindowState();
-		if (count < std::numeric_limits<uint32_t>::max()) {
-			window_state.qst32 = QuantileSortTree<uint32_t>::WindowInit<INPUT_TYPE>(data, aggr_input_data, data_mask,
-			                                                                 filter_mask, count);
-		} else {
-			window_state.qst64 = QuantileSortTree<uint64_t>::WindowInit<INPUT_TYPE>(data, aggr_input_data, data_mask,
-			                                                                 filter_mask, count);
-		}
-	}
-
-	static idx_t FrameSize(const QuantileIncluded &included, const SubFrames &frames) {
-		//	Count the number of valid values
-		idx_t n = 0;
-		if (included.AllValid()) {
-			for (const auto &frame : frames) {
-				n += frame.end - frame.start;
-			}
-		} else {
-			//	NULLs or FILTERed values,
-			for (const auto &frame : frames) {
-				for (auto i = frame.start; i < frame.end; ++i) {
-					n += included(i);
-				}
-			}
-		}
-
-		return n;
-	}
-};
-
 template <class STATE, class INPUT_TYPE, class RESULT_TYPE, class OP>
 static AggregateFunction QuantileListAggregate(const LogicalType &input_type, const LogicalType &child_type) { // NOLINT
 	LogicalType result_type =
@@ -940,218 +592,9 @@ AggregateFunction GetContinuousQuantileListAggregateFunction(const LogicalType &
 	}
 }
 
-//===--------------------------------------------------------------------===//
-// Median Absolute Deviation
-//===--------------------------------------------------------------------===//
-template <typename T, typename R, typename MEDIAN_TYPE>
-struct MadAccessor {
-	using INPUT_TYPE = T;
-	using RESULT_TYPE = R;
-	const MEDIAN_TYPE &median;
-	explicit MadAccessor(const MEDIAN_TYPE &median_p) : median(median_p) {
-	}
-
-	inline RESULT_TYPE operator()(const INPUT_TYPE &input) const {
-		const RESULT_TYPE delta = input - UnsafeNumericCast<RESULT_TYPE>(median);
-		return TryAbsOperator::Operation<RESULT_TYPE, RESULT_TYPE>(delta);
-	}
-};
-
-// hugeint_t - double => undefined
-template <>
-struct MadAccessor<hugeint_t, double, double> {
-	using INPUT_TYPE = hugeint_t;
-	using RESULT_TYPE = double;
-	using MEDIAN_TYPE = double;
-	const MEDIAN_TYPE &median;
-	explicit MadAccessor(const MEDIAN_TYPE &median_p) : median(median_p) {
-	}
-	inline RESULT_TYPE operator()(const INPUT_TYPE &input) const {
-		const auto delta = Hugeint::Cast<double>(input) - median;
-		return TryAbsOperator::Operation<double, double>(delta);
-	}
-};
-
-// date_t - timestamp_t => interval_t
-template <>
-struct MadAccessor<date_t, interval_t, timestamp_t> {
-	using INPUT_TYPE = date_t;
-	using RESULT_TYPE = interval_t;
-	using MEDIAN_TYPE = timestamp_t;
-	const MEDIAN_TYPE &median;
-	explicit MadAccessor(const MEDIAN_TYPE &median_p) : median(median_p) {
-	}
-	inline RESULT_TYPE operator()(const INPUT_TYPE &input) const {
-		const auto dt = Cast::Operation<date_t, timestamp_t>(input);
-		const auto delta = dt - median;
-		return Interval::FromMicro(TryAbsOperator::Operation<int64_t, int64_t>(delta));
-	}
-};
-
-// timestamp_t - timestamp_t => int64_t
-template <>
-struct MadAccessor<timestamp_t, interval_t, timestamp_t> {
-	using INPUT_TYPE = timestamp_t;
-	using RESULT_TYPE = interval_t;
-	using MEDIAN_TYPE = timestamp_t;
-	const MEDIAN_TYPE &median;
-	explicit MadAccessor(const MEDIAN_TYPE &median_p) : median(median_p) {
-	}
-	inline RESULT_TYPE operator()(const INPUT_TYPE &input) const {
-		const auto delta = input - median;
-		return Interval::FromMicro(TryAbsOperator::Operation<int64_t, int64_t>(delta));
-	}
-};
-
-// dtime_t - dtime_t => int64_t
-template <>
-struct MadAccessor<dtime_t, interval_t, dtime_t> {
-	using INPUT_TYPE = dtime_t;
-	using RESULT_TYPE = interval_t;
-	using MEDIAN_TYPE = dtime_t;
-	const MEDIAN_TYPE &median;
-	explicit MadAccessor(const MEDIAN_TYPE &median_p) : median(median_p) {
-	}
-	inline RESULT_TYPE operator()(const INPUT_TYPE &input) const {
-		const auto delta = input - median;
-		return Interval::FromMicro(TryAbsOperator::Operation<int64_t, int64_t>(delta));
-	}
-};
-
-template <typename MEDIAN_TYPE>
-struct MedianAbsoluteDeviationOperation : QuantileOperation {
-	template <class T, class STATE>
-	static void Finalize(STATE &state, T &target, AggregateFinalizeData &finalize_data) {
-		if (state.v.empty()) {
-			finalize_data.ReturnNull();
-			return;
-		}
-		using SAVE_TYPE = typename STATE::SaveType;
-		D_ASSERT(finalize_data.input.bind_data);
-		auto &bind_data = finalize_data.input.bind_data->Cast<QuantileBindData>();
-		D_ASSERT(bind_data.quantiles.size() == 1);
-		const auto &q = bind_data.quantiles[0];
-		Interpolator<false> interp(q, state.v.size(), false);
-		const auto med = interp.template Operation<SAVE_TYPE, MEDIAN_TYPE>(state.v.data(), finalize_data.result);
-
-		MadAccessor<SAVE_TYPE, T, MEDIAN_TYPE> accessor(med);
-		target = interp.template Operation<SAVE_TYPE, T>(state.v.data(), finalize_data.result, accessor);
-	}
-
-	template <class STATE, class INPUT_TYPE, class RESULT_TYPE>
-	static void Window(const INPUT_TYPE *data, const ValidityMask &fmask, const ValidityMask &dmask,
-	                   AggregateInputData &aggr_input_data, STATE &state, const SubFrames &frames, Vector &result,
-	                   idx_t ridx, const STATE *gstate) {
-		auto rdata = FlatVector::GetData<RESULT_TYPE>(result);
-
-		QuantileIncluded included(fmask, dmask);
-		const auto n = FrameSize(included, frames);
-
-		if (!n) {
-			auto &rmask = FlatVector::Validity(result);
-			rmask.Set(ridx, false);
-			return;
-		}
-
-		//	Compute the median
-		D_ASSERT(aggr_input_data.bind_data);
-		auto &bind_data = aggr_input_data.bind_data->Cast<QuantileBindData>();
-
-		D_ASSERT(bind_data.quantiles.size() == 1);
-		const auto &quantile = bind_data.quantiles[0];
-		auto &window_state = state.GetOrCreateWindowState();
-		MEDIAN_TYPE med;
-		if (gstate && gstate->HasTrees()) {
-			med = gstate->GetWindowState().template WindowScalar<MEDIAN_TYPE, false>(data, frames, n, result, quantile);
-		} else {
-			window_state.UpdateSkip(data, frames, included);
-			med = window_state.template WindowScalar<MEDIAN_TYPE, false>(data, frames, n, result, quantile);
-		}
-
-		//  Lazily initialise frame state
-		window_state.SetCount(frames.back().end - frames.front().start);
-		auto index2 = window_state.m.data();
-		D_ASSERT(index2);
-
-		// The replacement trick does not work on the second index because if
-		// the median has changed, the previous order is not correct.
-		// It is probably close, however, and so reuse is helpful.
-		auto &prevs = window_state.prevs;
-		ReuseIndexes(index2, frames, prevs);
-		std::partition(index2, index2 + window_state.count, included);
-
-		Interpolator<false> interp(quantile, n, false);
-
-		// Compute mad from the second index
-		using ID = QuantileIndirect<INPUT_TYPE>;
-		ID indirect(data);
-
-		using MAD = MadAccessor<INPUT_TYPE, RESULT_TYPE, MEDIAN_TYPE>;
-		MAD mad(med);
-
-		using MadIndirect = QuantileComposed<MAD, ID>;
-		MadIndirect mad_indirect(mad, indirect);
-		rdata[ridx] = interp.template Operation<idx_t, RESULT_TYPE, MadIndirect>(index2, result, mad_indirect);
-
-		//	Prev is used by both skip lists and increments
-		prevs = frames;
-	}
-};
-
 unique_ptr<FunctionData> BindMedian(ClientContext &context, AggregateFunction &function,
                                     vector<unique_ptr<Expression>> &arguments) {
 	return make_uniq<QuantileBindData>(Value::DECIMAL(int16_t(5), 2, 1));
-}
-
-template <typename INPUT_TYPE, typename MEDIAN_TYPE, typename TARGET_TYPE>
-AggregateFunction GetTypedMedianAbsoluteDeviationAggregateFunction(const LogicalType &input_type,
-                                                                   const LogicalType &target_type) {
-	using STATE = QuantileState<INPUT_TYPE, INPUT_TYPE>;
-	using OP = MedianAbsoluteDeviationOperation<MEDIAN_TYPE>;
-	auto fun = AggregateFunction::UnaryAggregateDestructor<STATE, INPUT_TYPE, TARGET_TYPE, OP>(input_type, target_type);
-	fun.bind = BindMedian;
-	fun.order_dependent = AggregateOrderDependent::NOT_ORDER_DEPENDENT;
-	fun.window = AggregateFunction::UnaryWindow<STATE, INPUT_TYPE, TARGET_TYPE, OP>;
-	fun.window_init = OP::template WindowInit<STATE, INPUT_TYPE>;
-	return fun;
-}
-
-AggregateFunction GetMedianAbsoluteDeviationAggregateFunction(const LogicalType &type) {
-	switch (type.id()) {
-	case LogicalTypeId::FLOAT:
-		return GetTypedMedianAbsoluteDeviationAggregateFunction<float, float, float>(type, type);
-	case LogicalTypeId::DOUBLE:
-		return GetTypedMedianAbsoluteDeviationAggregateFunction<double, double, double>(type, type);
-	case LogicalTypeId::DECIMAL:
-		switch (type.InternalType()) {
-		case PhysicalType::INT16:
-			return GetTypedMedianAbsoluteDeviationAggregateFunction<int16_t, int16_t, int16_t>(type, type);
-		case PhysicalType::INT32:
-			return GetTypedMedianAbsoluteDeviationAggregateFunction<int32_t, int32_t, int32_t>(type, type);
-		case PhysicalType::INT64:
-			return GetTypedMedianAbsoluteDeviationAggregateFunction<int64_t, int64_t, int64_t>(type, type);
-		case PhysicalType::INT128:
-			return GetTypedMedianAbsoluteDeviationAggregateFunction<hugeint_t, hugeint_t, hugeint_t>(type, type);
-		default:
-			throw NotImplementedException("Unimplemented Median Absolute Deviation DECIMAL aggregate");
-		}
-		break;
-
-	case LogicalTypeId::DATE:
-		return GetTypedMedianAbsoluteDeviationAggregateFunction<date_t, timestamp_t, interval_t>(type,
-		                                                                                         LogicalType::INTERVAL);
-	case LogicalTypeId::TIMESTAMP:
-	case LogicalTypeId::TIMESTAMP_TZ:
-		return GetTypedMedianAbsoluteDeviationAggregateFunction<timestamp_t, timestamp_t, interval_t>(
-		    type, LogicalType::INTERVAL);
-	case LogicalTypeId::TIME:
-	case LogicalTypeId::TIME_TZ:
-		return GetTypedMedianAbsoluteDeviationAggregateFunction<dtime_t, dtime_t, interval_t>(type,
-		                                                                                      LogicalType::INTERVAL);
-
-	default:
-		throw NotImplementedException("Unimplemented Median Absolute Deviation aggregate");
-	}
 }
 
 unique_ptr<FunctionData> BindMedianDecimal(ClientContext &context, AggregateFunction &function,
@@ -1164,14 +607,6 @@ unique_ptr<FunctionData> BindMedianDecimal(ClientContext &context, AggregateFunc
 	function.deserialize = QuantileBindData::Deserialize;
 	function.order_dependent = AggregateOrderDependent::NOT_ORDER_DEPENDENT;
 	return bind_data;
-}
-
-unique_ptr<FunctionData> BindMedianAbsoluteDeviationDecimal(ClientContext &context, AggregateFunction &function,
-                                                            vector<unique_ptr<Expression>> &arguments) {
-	function = GetMedianAbsoluteDeviationAggregateFunction(arguments[0]->return_type);
-	function.name = "mad";
-	function.order_dependent = AggregateOrderDependent::NOT_ORDER_DEPENDENT;
-	return BindMedian(context, function, arguments);
 }
 
 static const Value &CheckQuantile(const Value &quantile_val) {
@@ -1404,20 +839,6 @@ AggregateFunctionSet QuantileContFun::GetFunctions() {
 		}
 	}
 	return quantile_cont;
-}
-
-AggregateFunctionSet MadFun::GetFunctions() {
-	AggregateFunctionSet mad("mad");
-	mad.AddFunction(AggregateFunction({LogicalTypeId::DECIMAL}, LogicalTypeId::DECIMAL, nullptr, nullptr, nullptr,
-	                                  nullptr, nullptr, nullptr, BindMedianAbsoluteDeviationDecimal));
-
-	const vector<LogicalType> MAD_TYPES = {LogicalType::FLOAT,     LogicalType::DOUBLE, LogicalType::DATE,
-	                                       LogicalType::TIMESTAMP, LogicalType::TIME,   LogicalType::TIMESTAMP_TZ,
-	                                       LogicalType::TIME_TZ};
-	for (const auto &type : MAD_TYPES) {
-		mad.AddFunction(GetMedianAbsoluteDeviationAggregateFunction(type));
-	}
-	return mad;
 }
 
 } // namespace duckdb
