+++ src/include/duckdb/core_functions/aggregate/quantile_helpers.hpp
@@ -71,5 +71,4 @@ struct QuantileBindData : public FunctionData {
 	bool desc;
 };
 
-
 } // namespace duckdb
